requirements.txt
Django==3.2.3
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2==2.8.6
django-cities-light==3.9.0
django-filter==2.4.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
django-cors-headers==3.7.0
django-countries==7.2.1
django-ses==2.0.0
pycountry==20.7.3
PyYAML==5.4.1
ua-parser==0.10.0
user-agents==2.2.0
django-user-agents==0.4.0
pdfkit==0.6.1
Jinja2==3.0.1
docusign-esign==3.10.0
cryptography==3.4.7
requests==2.25.1
django-apscheduler==0.6.0
pandas==1.1.5
openpyxl==3.0.9
XlsxWriter==3.0.2
Pillow==8.4.0
python-decouple==3.6